Fix toggle dark mode button
