+++ src/styles/style.scss
@@ -54,6 +54,10 @@ header {
       justify-content: center;
       align-items: center;
       cursor: pointer;
+
+      &:hover {
+        background: rgb(205, 232, 250);
+      }
     }
 
     button {
